pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "homology"
version = "0.1.0"
description = "Identify structural holes in Obsidian knowledge graphs using embeddings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "networkx>=3.0",
    "scikit-learn>=1.3",
    "openai>=1.0",
    "tiktoken>=0.5",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
local = [
    "sentence-transformers>=2.2",
]

[project.scripts]
homology = "homology.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
